backend/ai-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
asyncpg==0.29.0
redis==5.0.1
pydantic==2.5.3
python-dotenv==1.0.0
openai==1.7.2
langchain==0.1.0
langchain-openai==0.0.2
ortools==9.8.3296
numpy==1.26.3
pytest==7.4.4
pytest-asyncio==0.23.3
python-jose[cryptography]==3.3.0
